refactor: decompose ProfileSection and FlowDiagram into folders
큰 파일 두 개를 추가로 폴더 구조로 분해. 동작 변경 없음.

### ProfileSection (874 → 169줄 메인 + 11 파일)
src/sections/home/ProfileSection.tsx → ProfileSection/index.tsx (이동, git mv)
- types.ts (TabId, HighlightItem, AiHighlightItem, WorkItem, EducationItem, AwardItem, CertItem, SkillGroup, ProfileData)
- animations.ts (contentVariants, listVariants, itemVariants, resumeLinkOrder)
- prefetch.ts (cachedData + dataPromise — 모듈 로드 시점 fetch)
- SkillsBlock.tsx
- ResumeProfileCard.tsx
- BasicsTab.tsx / WorkSkillsTab.tsx / EducationTab.tsx / AwardsAndCertsTab.tsx
- WorkBlock/ (252줄 — Dev / AI 하이라이트 아코디언; 별도 폴더)
  - index.tsx (상태 + 메인 JSX)
  - HighlightCard.tsx (3-step 타임라인)
  - AiHighlightCard.tsx (4-step 타임라인 + nested toggle)

### FlowDiagram (671 → 397줄 메인 + 7 파일)
src/components/AiDevKitModal/FlowDiagram.tsx → FlowDiagram/index.tsx
- constants.ts (STEP_WIDTH, LOOP_BLOCK_WIDTH 등 레이아웃 상수, readableTextStyle)
- types.ts (FlowStep, FlowBlock, FlowDiagramProps)
- layout.ts (toStepItems, sanitizeId, getArrowPath, getLoopNodePositions — 순수 함수)
- DownArrow.tsx (모바일용 세로 화살표)
- FlowStepCard.tsx (메인 라인 카드)
- LoopStepCard.tsx (루프 클러스터 내부 작은 카드)
- LoopLabels.tsx (루프 의미 설명 라벨)

### 테스트
- FlowDiagram/layout.test.ts — toStepItems / sanitizeId / getArrowPath / getLoopNodePositions 12개 테스트 신규 추가
- 전체 302개 테스트 통과
- TypeScript 컴파일 + 프로덕션 빌드 정상
